Recognition fix

+++ NeuralNets/FaceRecognition/Recognition.py
@@ -2,7 +2,6 @@ import os
 from threading import Thread
 
 import face_recognition
-import matplotlib.pyplot as plt
 import numpy as np
 
 import cv2
@@ -13,18 +12,19 @@ x_placeholder = None
 model_probs = None
 face_detector = None
 
-names = []
-person_faces_amount = 150
-neural_net_input_shape = [300, 300]
-n_iters_predict = 500
-n_iters_train = 14
-predict_threshold = 0.85
-batch_size = 32
 model_path = "./NeuralNets/FaceRecognition/fc_model.ckpt"
 encodings_path = "./NeuralNets/FaceRecognition/encodings"
 dataset_path = "/home/muroming/PythonProjects/SmartHouse/NeuralNets/FaceRecognition/dataset"
 face_cv_path = "./NeuralNets/FaceRecognition/%s"
 encodings_name = "%s.npy"
+
+names = []
+person_faces_amount = 150
+batch_size = 32
+neural_net_input_shape = [300, 300]
+n_iters_predict = 500
+n_iters_train = int(person_faces_amount / batch_size * 3)
+predict_threshold = 0.65
 
 
 class TrainingThread(Thread):
@@ -34,11 +34,11 @@ class TrainingThread(Thread):
         self.run()
 
     def run(self):
-        # try:
-        train_classifier()
-        # except Exception as e:
-        #     print("Training failed with exception")
-        #     print(e)
+        try:
+            train_classifier()
+        except Exception as e:
+            print("Training failed with exception")
+            print(e)
 
 
 def load_model():
@@ -128,14 +128,12 @@ def save_person_encodings(encodings, name):
     np.save(encodings, os.path.join(encodings_path, encodings_name % name))
 
 
-# def load_names():
-#     global names
-#     names = [name[:-4] for name in os.listdir(encodings_path)]
-#     print("Names loaded", names)
-
 def load_names():
     global names
-    names = ["nik", "tan", "mock"]  # os.listdir(dataset_path)
+    removes = ["testfaces"]
+    names = [name for name in os.listdir(
+        dataset_path) if name not in removes and not ".npy" in name]
+
     print("Names loaded", names)
 
 
@@ -158,68 +156,40 @@ def train_classifier():
         return
 
     print("Getting images")
-    # X, y = np.load(dataset_path + "/X.npy")[:person_faces_amount], np.load(dataset_path + "/y.npy")[:person_faces_amount]
-    X, y = [], []
-    persons = 0
-    for person in names:
-        person_path = os.path.join(dataset_path, person)
-        for i, image in enumerate(os.listdir(person_path)):
-            print(person, i)
-            if i == person_faces_amount:
-                break
-            img = cv2.imread(os.path.join(person_path, image))
-            img = cv2.cvtColor(img, cv2.COLOR_BGR2RGB)
-            enc = face_recognition.face_encodings(img, [(0, img.shape[1], img.shape[0], 0)])
-            # img = img / 255.
-            # img = cv2.resize(img, tuple(neural_net_input_shape))
-            enc = np.array(enc).reshape((128,))
-            X.append(enc)
-            y.append(persons)
+    x_path = os.path.join(dataset_path, "X.npy")
+    y_path = os.path.join(dataset_path, "y.npy")
 
-        persons += 1
-    print("Preparing targets")
-    X = np.array(X)
-    y = person_id_to_bin(y)
+    if os.path.exists(x_path) and os.path.exists(y_path):
+        X, y = np.load(dataset_path + "/X.npy"), np.load(dataset_path + "/y.npy")
+    else:
+        X, y = [], []
+        persons = 0
+        for person in names:
+            person_path = os.path.join(dataset_path, person)
+            for i, image in enumerate(os.listdir(person_path)):
+                print(person, i)
+                if i == person_faces_amount:
+                    break
+                img = cv2.imread(os.path.join(person_path, image))
+                img = cv2.cvtColor(img, cv2.COLOR_BGR2RGB)
+                enc = face_recognition.face_encodings(img, [(0, img.shape[1], img.shape[0], 0)])
+                enc = np.array(enc).reshape((128,))
+                X.append(enc)
+                y.append(persons)
 
-    np.save(dataset_path + "/X.npy", X)
-    np.save(dataset_path + "/y.npy", y)
+            persons += 1
+        print("Preparing targets")
+        X = np.array(X)
+        y = person_id_to_bin(y)
+
+        np.save(dataset_path + "/X.npy", X)
+        np.save(dataset_path + "/y.npy", y)
 
     print(X.shape, y.shape)
     print(y)
 
     print("Training model")
     train_model(len(names), X, y)
-
-
-# def train_classifier():
-#     if len(names) == 0:
-#         print("Loading names")
-#         load_names()
-#
-#     if len(names) <= 1:
-#         print("Can't train on %d person" % len(names))
-#         return
-#
-#     print("Getting encodings")
-#     X, y = [], []
-#     persons = 0
-#     for person in os.listdir(encodings_path):
-#         name = person[:-4]
-#         print("Encodings for %s" % name)
-#         enc = np.load(os.path.join(encodings_path, person))
-#         if len(X) != 0:
-#             X = np.vstack((X, enc))
-#         else:
-#             X = enc
-#         names[persons] = name
-#         y.extend([persons for _ in range(enc.shape[0])])
-#         persons += 1
-#
-#     print("Preparing targets")
-#     y = person_id_to_bin(y)
-#
-#     print("Training model")
-#     train_model(persons, X, y)
 
 
 def validate_person(image, detection_type=0):  # Assuming image is RGB
@@ -240,12 +210,12 @@ def validate_person(image, detection_type=0):  # Assuming image is RGB
 
 
 load_model()
-
-img = cv2.imread(os.path.join(dataset_path, "nik", "nik973rot.jpg"))
-enc = face_recognition.face_encodings(img, [(0, img.shape[1], img.shape[0], 0)])
-enc = np.array(enc).reshape((128,))
-model.predict([enc])
-
-img = cv2.imread(os.path.join(dataset_path, "testfaces", "suprised-man.jpg"))
-enc = extract_face_enc_from_image(img)
-model.predict(enc)
+#
+# img = cv2.imread(os.path.join(dataset_path, "nik", "nik935rot.jpg"))
+# enc = face_recognition.face_encodings(img, [(0, img.shape[1], img.shape[0], 0)])
+# enc = np.array(enc).reshape((128,))
+# model.predict([enc])
+#
+# img = cv2.imread(os.path.join(dataset_path, "mock", "mock1.jpg"))
+# enc = extract_face_enc_from_image(img)
+# model.predict(enc)
